+++ changelog.md
@@ -7,6 +7,9 @@
  - On mobile screen sizes, popovers now have a margin to prevent them from filling the entire width of their container.
 
 **Fixed:**
+- bpk-component-banner-alert:
+  - Fixed issue where banner was shown if `show` was initially set `false`.
+
 - bpk-component-popover:
 - bpk-react-utils:
   - `target` prop type is now correctly `oneOf([function, node])` instead of `oneOf([function, element])`